Add Jupyter Notebook example (#206)
This PR adds an Jupyter Notebook to the examples folder (capturing some
learnings from the Knowledge Graph Conference).. Whilst testing a Dgraph
Cloud version of the Notebook in this PR, I discovered a bug in the
client_stub module whereby if both the protocol scheme (https) and the
port number (:443) were included in the call to `from_cloud`, the
connection would fail. This format was specifically illustrated in the
README, so it was worth fixing. Although I did correct the example on
the README to be the grpc endpoint value that is presented on the Dgraph
Cloud Settings tab.

I'll leave it to you @joshua-goldstein to whether you want to bump the
official version and push to PyPI.

---------

Co-authored-by: Joshua Goldstein <92491720+joshua-goldstein@users.noreply.github.com>